FEAT: 메뉴 기능 변경
- 메뉴 라우팅 fullScreenModal로 방식 변경
- 메뉴 라우팅 방식 dismiss로 변경
