+++ app/_layout.tsx
@@ -1,3 +1,4 @@
+import MenuHeader from '@/components/organisms/headers/MenuHeader';
 import { useReactQueryDevTools } from '@dev-plugins/react-query';
 import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 import { DefaultTheme, ThemeProvider } from '@react-navigation/native';
@@ -62,7 +63,16 @@ export default function RootLayout() {
           <BottomSheetModalProvider>
             <SafeAreaProvider>
               <StatusBar style="dark" />
-              <Stack />
+              <Stack>
+                <Stack.Screen name="index" options={{ headerShown: false }} />
+                <Stack.Screen
+                  name="menu"
+                  options={{
+                    header: () => <MenuHeader />,
+                    presentation: 'fullScreenModal'
+                  }}
+                />
+              </Stack>
             </SafeAreaProvider>
           </BottomSheetModalProvider>
         </GestureHandlerRootView>
